pyproject.toml
[project]
name = "multi-llm"
version = "1.0.0"
description = "Multi-LLM Fan-Out & Evaluation System"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
]
